Move Sky(RD) into environment Move Fog logic from render scene render to fog
