+++ servers/rendering/renderer_rd/renderer_scene_render_rd.h
@@ -40,11 +40,10 @@
 #include "servers/rendering/renderer_rd/effects/ss_effects.h"
 #include "servers/rendering/renderer_rd/effects/tone_mapper.h"
 #include "servers/rendering/renderer_rd/effects/vrs.h"
+#include "servers/rendering/renderer_rd/environment/fog.h"
 #include "servers/rendering/renderer_rd/environment/gi.h"
+#include "servers/rendering/renderer_rd/environment/sky.h"
 #include "servers/rendering/renderer_rd/renderer_scene_environment_rd.h"
-#include "servers/rendering/renderer_rd/renderer_scene_sky_rd.h"
-#include "servers/rendering/renderer_rd/shaders/volumetric_fog.glsl.gen.h"
-#include "servers/rendering/renderer_rd/shaders/volumetric_fog_process.glsl.gen.h"
 #include "servers/rendering/renderer_scene.h"
 #include "servers/rendering/renderer_scene_render.h"
 #include "servers/rendering/rendering_device.h"
@@ -99,7 +98,7 @@ struct RenderDataRD {
 };
 
 class RendererSceneRenderRD : public RendererSceneRender {
-	friend RendererSceneSkyRD;
+	friend RendererRD::SkyRD;
 	friend RendererRD::GI;
 
 protected:
@@ -167,7 +166,7 @@ protected:
 
 	RendererRD::SSEffects *ss_effects = nullptr;
 	RendererRD::GI gi;
-	RendererSceneSkyRD sky;
+	RendererRD::SkyRD sky;
 
 	RendererSceneEnvironmentRD *get_environment(RID p_environment) {
 		if (p_environment.is_valid()) {
@@ -212,7 +211,7 @@ private:
 
 		struct Reflection {
 			RID owner;
-			RendererSceneSkyRD::ReflectionData data;
+			RendererRD::SkyRD::ReflectionData data;
 			RID fbs[6];
 		};
 
@@ -407,16 +406,6 @@ private:
 
 	mutable RID_Owner<LightInstance> light_instance_owner;
 
-	/* FOG VOLUMES */
-
-	struct FogVolumeInstance {
-		RID volume;
-		Transform3D transform;
-		bool active = false;
-	};
-
-	mutable RID_Owner<FogVolumeInstance> fog_volume_instance_owner;
-
 	/* ENVIRONMENT */
 
 	RS::EnvironmentSSAOQuality ssao_quality = RS::ENV_SSAO_QUALITY_MEDIUM;
@@ -471,8 +460,6 @@ private:
 	ClusterBuilderSharedDataRD cluster_builder_shared;
 	ClusterBuilderRD *current_cluster_builder = nullptr;
 
-	struct VolumetricFog;
-
 	struct RenderBuffers {
 		RenderBufferData *data = nullptr;
 		int internal_width = 0;
@@ -508,8 +495,8 @@ private:
 		Vector<View> views;
 
 		RendererRD::GI::SDFGI *sdfgi = nullptr;
-		VolumetricFog *volumetric_fog = nullptr;
 		RendererRD::GI::RenderBuffersGI rbgi;
+		RendererRD::Fog::VolumetricFog *volumetric_fog = nullptr;
 
 		ClusterBuilderRD *cluster_builder = nullptr;
 
@@ -749,204 +736,6 @@ private:
 		bool depth_prepass_used; // this does not seem used anywhere...
 	} render_state;
 
-	struct VolumetricFog {
-		enum {
-			MAX_TEMPORAL_FRAMES = 16
-		};
-
-		uint32_t width = 0;
-		uint32_t height = 0;
-		uint32_t depth = 0;
-
-		float length;
-		float spread;
-
-		RID light_density_map;
-		RID prev_light_density_map;
-		RID fog_map;
-		RID density_map;
-		RID light_map;
-		RID emissive_map;
-
-		RID fog_uniform_set;
-		RID copy_uniform_set;
-		RID process_uniform_set_density;
-		RID process_uniform_set;
-		RID process_uniform_set2;
-		RID sdfgi_uniform_set;
-		RID sky_uniform_set;
-
-		int last_shadow_filter = -1;
-	};
-
-	struct VolumetricFogShader {
-		enum FogSet {
-			FOG_SET_BASE,
-			FOG_SET_UNIFORMS,
-			FOG_SET_MATERIAL,
-			FOG_SET_MAX,
-		};
-
-		struct FogPushConstant {
-			float position[3];
-			float pad;
-
-			float extents[3];
-			float pad2;
-
-			int32_t corner[3];
-			uint32_t shape;
-
-			float transform[16];
-		};
-
-		struct VolumeUBO {
-			float fog_frustum_size_begin[2];
-			float fog_frustum_size_end[2];
-
-			float fog_frustum_end;
-			float z_near;
-			float z_far;
-			float time;
-
-			int32_t fog_volume_size[3];
-			uint32_t directional_light_count;
-
-			uint32_t use_temporal_reprojection;
-			uint32_t temporal_frame;
-			float detail_spread;
-			float temporal_blend;
-
-			float to_prev_view[16];
-			float transform[16];
-		};
-
-		ShaderCompiler compiler;
-		VolumetricFogShaderRD shader;
-		FogPushConstant push_constant;
-		RID volume_ubo;
-
-		RID default_shader;
-		RID default_material;
-		RID default_shader_rd;
-
-		RID base_uniform_set;
-
-		RID params_ubo;
-
-		enum {
-			VOLUMETRIC_FOG_PROCESS_SHADER_DENSITY,
-			VOLUMETRIC_FOG_PROCESS_SHADER_DENSITY_WITH_SDFGI,
-			VOLUMETRIC_FOG_PROCESS_SHADER_FILTER,
-			VOLUMETRIC_FOG_PROCESS_SHADER_FOG,
-			VOLUMETRIC_FOG_PROCESS_SHADER_COPY,
-			VOLUMETRIC_FOG_PROCESS_SHADER_MAX,
-		};
-
-		struct ParamsUBO {
-			float fog_frustum_size_begin[2];
-			float fog_frustum_size_end[2];
-
-			float fog_frustum_end;
-			float ambient_inject;
-			float z_far;
-			uint32_t filter_axis;
-
-			float ambient_color[3];
-			float sky_contribution;
-
-			int32_t fog_volume_size[3];
-			uint32_t directional_light_count;
-
-			float base_emission[3];
-			float base_density;
-
-			float base_scattering[3];
-			float phase_g;
-
-			float detail_spread;
-			float gi_inject;
-			uint32_t max_voxel_gi_instances;
-			uint32_t cluster_type_size;
-
-			float screen_size[2];
-			uint32_t cluster_shift;
-			uint32_t cluster_width;
-
-			uint32_t max_cluster_element_count_div_32;
-			uint32_t use_temporal_reprojection;
-			uint32_t temporal_frame;
-			float temporal_blend;
-
-			float cam_rotation[12];
-			float to_prev_view[16];
-			float radiance_inverse_xform[12];
-		};
-
-		VolumetricFogProcessShaderRD process_shader;
-
-		RID process_shader_version;
-		RID process_pipelines[VOLUMETRIC_FOG_PROCESS_SHADER_MAX];
-
-	} volumetric_fog;
-
-	uint32_t volumetric_fog_depth = 128;
-	uint32_t volumetric_fog_size = 128;
-	bool volumetric_fog_filter_active = true;
-
-	Vector3i _point_get_position_in_froxel_volume(const Vector3 &p_point, float fog_end, const Vector2 &fog_near_size, const Vector2 &fog_far_size, float volumetric_fog_detail_spread, const Vector3 &fog_size, const Transform3D &p_cam_transform);
-	void _volumetric_fog_erase(RenderBuffers *rb);
-	void _update_volumetric_fog(RID p_render_buffers, RID p_environment, const Projection &p_cam_projection, const Transform3D &p_cam_transform, const Transform3D &p_prev_cam_inv_transform, RID p_shadow_atlas, int p_directional_light_count, bool p_use_directional_shadows, int p_positional_light_count, int p_voxel_gi_count, const PagedArray<RID> &p_fog_volumes);
-
-	struct FogShaderData : public RendererRD::ShaderData {
-		bool valid = false;
-		RID version;
-
-		RID pipeline;
-		HashMap<StringName, ShaderLanguage::ShaderNode::Uniform> uniforms;
-		Vector<ShaderCompiler::GeneratedCode::Texture> texture_uniforms;
-
-		Vector<uint32_t> ubo_offsets;
-		uint32_t ubo_size = 0;
-
-		String path;
-		String code;
-		HashMap<StringName, HashMap<int, RID>> default_texture_params;
-
-		bool uses_time = false;
-
-		virtual void set_code(const String &p_Code);
-		virtual void set_path_hint(const String &p_hint);
-		virtual void set_default_texture_param(const StringName &p_name, RID p_texture, int p_index);
-		virtual void get_param_list(List<PropertyInfo> *p_param_list) const;
-		virtual void get_instance_param_list(List<RendererMaterialStorage::InstanceShaderParam> *p_param_list) const;
-		virtual bool is_param_texture(const StringName &p_param) const;
-		virtual bool is_animated() const;
-		virtual bool casts_shadows() const;
-		virtual Variant get_default_parameter(const StringName &p_parameter) const;
-		virtual RS::ShaderNativeSourceCode get_native_source_code() const;
-
-		FogShaderData() {}
-		virtual ~FogShaderData();
-	};
-
-	struct FogMaterialData : public RendererRD::MaterialData {
-		FogShaderData *shader_data = nullptr;
-		RID uniform_set;
-		bool uniform_set_updated;
-
-		virtual void set_render_priority(int p_priority) {}
-		virtual void set_next_pass(RID p_pass) {}
-		virtual bool update_parameters(const HashMap<StringName, Variant> &p_parameters, bool p_uniform_dirty, bool p_textures_dirty);
-		virtual ~FogMaterialData();
-	};
-
-	RendererRD::ShaderData *_create_fog_shader_func();
-	static RendererRD::ShaderData *_create_fog_shader_funcs();
-
-	RendererRD::MaterialData *_create_fog_material_func(FogShaderData *p_shader);
-	static RendererRD::MaterialData *_create_fog_material_funcs(RendererRD::ShaderData *p_shader);
-
 	RID shadow_sampler;
 
 	uint64_t scene_pass = 0;
@@ -962,6 +751,14 @@ private:
 	uint32_t max_cluster_elements = 512;
 
 	void _render_shadow_pass(RID p_light, RID p_shadow_atlas, int p_pass, const PagedArray<GeometryInstance *> &p_instances, const Plane &p_camera_plane = Plane(), float p_lod_distance_multiplier = 0, float p_screen_mesh_lod_threshold = 0.0, bool p_open_pass = true, bool p_close_pass = true, bool p_clear_region = true, RendererScene::RenderInfo *p_render_info = nullptr);
+
+	/* Volumetric Fog */
+
+	uint32_t volumetric_fog_size = 128;
+	uint32_t volumetric_fog_depth = 128;
+	bool volumetric_fog_filter_active = true;
+
+	void _update_volumetric_fog(RID p_render_buffers, RID p_environment, const Projection &p_cam_projection, const Transform3D &p_cam_transform, const Transform3D &p_prev_cam_inv_transform, RID p_shadow_atlas, int p_directional_light_count, bool p_use_directional_shadows, int p_positional_light_count, int p_voxel_gi_count, const PagedArray<RID> &p_fog_volumes);
 
 public:
 	virtual Transform3D geometry_instance_get_transform(GeometryInstance *p_instance) = 0;
