+++ servers/rendering/renderer_rd/shaders/environment/sky.glsl
